config: support external configurations
It is sometimes convenient to have more than one configuration, either as
an alternative or as a parallel builder.  Support the ability to move
critical configuration components into a separate directory tree while
preserving the core yocto-autobuilder clone in a clean state.

The workflow is unchanged for the common use-case.  If you wish to have an
external (to the yocto-autobuilder tree) configuration, you may define
YOCTO_AB_CONFIG_DIR in your environment prior to sourcing
yocto-autobuilder-setup.  Something like this would work:

    $ YOCTO_AB_CONFIG_DIR=/build/ab-new . ./yocto-autobuilder-setup

YOCTO_AB_CONFIG_DIR need not exist prior to sourcing the setup script, but
the parent directory must be writable.  If YOCTO_AB_CONFIG_DIR does
already exist, the setup script will attempt to re-use that directory.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>